some comps moved to separate files, minor changes done
