+++ src/pages/NotFound/NotFound.tsx
@@ -1,0 +1,5 @@
+export const NotFound = () => {
+  return (
+    <h1 className="title">Page not found</h1>
+  );
+};
